Refactoring model packages
