+++ app/src/main/java/com/do55anto5/movieapp/domain/local/usecase/GetMoviesUseCase.kt
@@ -2,7 +2,7 @@ package com.do55anto5.movieapp.domain.local.usecase
 
 import com.do55anto5.movieapp.data.mapper.toDomain
 import com.do55anto5.movieapp.domain.local.repository.MovieLocalRepository
-import com.do55anto5.movieapp.domain.model.Movie
+import com.do55anto5.movieapp.domain.model.movie.Movie
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
