+++ app/src/main/java/com/do55anto5/movieapp/domain/local/usecase/InsertMovieUseCase.kt
@@ -2,7 +2,7 @@ package com.do55anto5.movieapp.domain.local.usecase
 
 import com.do55anto5.movieapp.data.mapper.toEntity
 import com.do55anto5.movieapp.domain.local.repository.MovieLocalRepository
-import com.do55anto5.movieapp.domain.model.Movie
+import com.do55anto5.movieapp.domain.model.movie.Movie
 import javax.inject.Inject
 
 class InsertMovieUseCase @Inject constructor(
