+++ app/src/main/java/com/do55anto5/movieapp/domain/model/movie/MovieReview.kt
@@ -1,4 +1,4 @@
-package com.do55anto5.movieapp.domain.model
+package com.do55anto5.movieapp.domain.model.movie
 
 import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
