+++ app/src/main/java/com/do55anto5/movieapp/domain/repository/movie/MovieRepository.kt
@@ -1,9 +1,9 @@
 package com.do55anto5.movieapp.domain.repository.movie
 
 import androidx.paging.PagingSource
-import com.do55anto5.movieapp.data.model.GenresResponse
-import com.do55anto5.movieapp.data.model.MovieResponse
-import com.do55anto5.movieapp.data.model.RemoteBasePagination
+import com.do55anto5.movieapp.data.model.movie.GenresResponse
+import com.do55anto5.movieapp.data.model.movie.MovieResponse
+import com.do55anto5.movieapp.data.model.movie.RemoteBasePagination
 
 interface MovieRepository {
