+++ app/src/main/java/com/do55anto5/movieapp/domain/usecase/movie/GetGenresUseCase.kt
@@ -1,7 +1,7 @@
 package com.do55anto5.movieapp.domain.usecase.movie
 
 import com.do55anto5.movieapp.data.mapper.toDomain
-import com.do55anto5.movieapp.domain.model.Genre
+import com.do55anto5.movieapp.domain.model.movie.Genre
 import com.do55anto5.movieapp.domain.repository.movie.MovieRepository
 import javax.inject.Inject
