+++ app/src/main/java/com/do55anto5/movieapp/domain/usecase/movie/SearchMoviesUseCase.kt
@@ -5,7 +5,7 @@ import androidx.paging.PagingConfig
 import androidx.paging.PagingData
 import androidx.paging.map
 import com.do55anto5.movieapp.data.mapper.toDomain
-import com.do55anto5.movieapp.domain.model.Movie
+import com.do55anto5.movieapp.domain.model.movie.Movie
 import com.do55anto5.movieapp.domain.repository.movie.MovieRepository
 import com.do55anto5.movieapp.util.Constants.Paging.DEFAULT_PAGE_INDEX
 import com.do55anto5.movieapp.util.Constants.Paging.NETWORK_PAGE_SIZE
