+++ app/src/main/java/com/do55anto5/movieapp/presenter/main/bottom_bar/download/DownloadViewModel.kt
@@ -6,7 +6,7 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.do55anto5.movieapp.domain.local.usecase.DeleteMovieUseCase
 import com.do55anto5.movieapp.domain.local.usecase.GetMoviesUseCase
-import com.do55anto5.movieapp.domain.model.Movie
+import com.do55anto5.movieapp.domain.model.movie.Movie
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.launch
 import javax.inject.Inject
