+++ app/src/main/java/com/do55anto5/movieapp/presenter/main/bottom_bar/home/adapter/MovieAdapter.kt
@@ -10,7 +10,7 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.do55anto5.movieapp.R
-import com.do55anto5.movieapp.domain.model.Movie
+import com.do55anto5.movieapp.domain.model.movie.Movie
 
 class MovieAdapter(
     private val context: Context,
